refactor delete commit operation

+++ src/controllers/base_controller.ts
@@ -1,6 +1,10 @@
 import { Request, Response } from "express";
 import { Model } from "mongoose";
 
+interface authenticatedRequest extends Request{
+    userId: string
+}
+
 abstract class BaseController<T> {
     model: Model<T>;
     constructor(model: any) {
@@ -44,10 +48,19 @@ abstract class BaseController<T> {
         }
     };
 
-    async delete(req: Request, res: Response) {
+    async delete(req: authenticatedRequest, res: Response) {
         try {
+            const authenticatedUserId = req.userId; // ID of the logged-in user
+            const ItemId = req.body.ownerId; // ID of the item to delete from the route
+
+            // Ensure that only the authenticated user can delete their Items
+            if (authenticatedUserId !== ItemId){
+                return res.status(403).send('Forbbiden');
+            }
+            
             const deletedItem = await this.model.findByIdAndDelete(req.params.id);
             if (!deletedItem) return res.status(404).json({ message: 'Not found' });
+
             res.status(200).json(deletedItem);
         } catch (err: any) {
             res.status(500).json({ error: err.message });
